+++ CHANGE_LOG.md
@@ -13,3 +13,4 @@
 - added color settings;
 - moved logs to file;
 - moved to World center-centered Millimeter based coordinate system;
+- user position is calculated based on the screen center position;